First version of cutflow tool

+++ source/EventObservables/src/EventObservablesVV.cc
@@ -128,9 +128,9 @@ void EventObservablesVV::updateChannelValues(EVENT::LCEvent *pLCEvent) {
     m_JMK_best = (m_fitchi2_ZHH <= m_fitchi2_ZZH ? m_JMK_ZHH : m_JMK_ZZH);
 
     streamlog_out(MESSAGE) << m_leps4cKinFit_4v.size() << " vs " << m_nAskedIsoLeps() << endl;
-    streamlog_out(MESSAGE) << (int)m_JMK_best.size() << " vs " << m_nAskedJets() << endl;
+    streamlog_out(MESSAGE) << m_JMK_best.size() << " vs " << m_nAskedJets() << endl;
       
-    if ((int)m_leps4cKinFit_4v.size() == m_nAskedIsoLeps() && (int)m_JMK_best.size() >= m_nAskedJets()) {
+    if (m_leps4cKinFit_4v.size() == m_nAskedIsoLeps() && m_JMK_best.size() >= m_nAskedJets()) {
       m_fit4C_masses.push_back((m_jets4cKinFit_4v[m_JMK_best[0]]+m_jets4cKinFit_4v[m_JMK_best[1]]).M());
       m_fit4C_masses.push_back((m_jets4cKinFit_4v[m_JMK_best[2]]+m_jets4cKinFit_4v[m_JMK_best[3]]).M());
       if (abs(m_fit4C_masses[0]-125.) < abs(m_fit4C_masses[1]-125.)) {
